+++ src/main/java/com/example/demo/Dtos/PostRequest.java
@@ -3,10 +3,12 @@ package com.example.demo.Dtos;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.time.LocalDateTime;
+
 @Getter
 @Setter
 
 public class PostRequest {
     private String content;
-    private Long authorId;
+    private int authorId;
 }
